pgseeder/pyproject.toml
[tool.poetry]
name = "thejitu-wat"
version = "0.1.0"
description = ""
authors = ["TruckTAR <57882685+trucktar@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "thejitu_wat"}]

[tool.poetry.dependencies]
python = "^3.9"
psycopg = {extras = ["c"], version = "^3.1.8"}
psycopg-pool = "^3.1.6"
openpyxl = "^3.1.1"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
